Type r.Lumen.ScreenProbeGather.Temporal.MaxFramesAccumulated as the float it is
The layout dump added in 0.1.6131 cracked the one CVar that had been reachable
but unreachable in Talos across every session. Its shadow at object+0x58 reads
0x41C80000, which is 25.0f: the variable is a float and the spec table declared
it Int32, so the Int32 plausibility check refused the reference layout every
time.

The refusal was right. Installing it as an int would have written 10 into a
float global - 1.4e-44, a denormal - and turned temporal accumulation off
entirely, which is the opposite of what the RR-optimal bundle is for.

The type checks stay fail-closed in the other direction too: a build that
registers it as an int presents 10 or 25, both of which reinterpret as denormal
floats and are refused rather than written. So the entry cannot silently corrupt
the variable it is meant to set, whichever way a title registers it.

Also records the show flag bit map measured in 20260816_201740: Bloom 1,
Tonemapper 3, AntiAliasing 4, TemporalAA 5, GlobalIllumination 12, Vignette 13,
Grain 14, DirectLighting 28, MotionBlur 37, SceneColorFringe 46, Lighting 109.
GlobalIllumination at 12 is exactly what the engine's own name table predicted,
one below the first bit 0.1.6128 set, so the console-object side is
self-consistent and the unproven part is how the renderer indexes the mask.

+++ hook/common/ue5_cvar_override_policy.h
@@ -67,7 +67,15 @@ inline constexpr std::array kSpecs{
          Activation::RayReconstructionOptimal, 1.0},
     Spec{"r.Lumen.ScreenProbeGather.SpatialFilterNumPasses", ValueType::Int32,
          Activation::RayReconstructionOptimal, 3.0},
-    Spec{"r.Lumen.ScreenProbeGather.Temporal.MaxFramesAccumulated", ValueType::Int32,
+    // Float, not int: Talos's console object carries a shadow of 25.0f
+    // (0x41C80000) behind its reference pointer, which the Int32 plausibility
+    // check refused for the whole life of this entry. The refusal was right -
+    // installing it as an int would have written 10 (1.4e-44 as a float) into a
+    // float global, i.e. no temporal accumulation at all. The type checks stay
+    // fail-closed in the other direction too: an int-typed build would present
+    // 10 or 25, which reinterpret as denormal floats and are refused rather
+    // than written.
+    Spec{"r.Lumen.ScreenProbeGather.Temporal.MaxFramesAccumulated", ValueType::Float,
          Activation::RayReconstructionOptimal, 10.0},
     Spec{"r.Lumen.ScreenProbeGather.Temporal.MaxRayDirections", ValueType::Int32,
          Activation::RayReconstructionOptimal, 8.0},
